Rename to mentat-lss
