+++ docs/api/emulator.rst
@@ -3,7 +3,7 @@
 Emulator API
 ============
 
-This page contains the main functions you need to interact with spherex_emu.
+This page contains the main functions you need to interact with mentat-lss.
 Specifically, these functions create and use the galaxy power spectrum emulator.
 
 .. automodule:: emulator
