+++ src/mentat_lss/fastpt/J_k.py
