+++ src/mentat_lss/fastpt/gamma_funcs.py
