+++ src/mentat_lss/models/stacked_mlp.py
@@ -3,10 +3,10 @@ import torch.nn as nn
 import math
 import itertools
 
-import spherex_emu.models.blocks as blocks
+import mentat_lss.models.blocks as blocks
 
-class single_transformer(nn.Module):
-    """Class defining a single independent transformer network"""
+class single_mlp(nn.Module):
+    """Class defining a single independent mlp network"""
 
     def __init__(self, config_dict, is_cross_spectra:bool):
         """Initializes an individual network, responsible for outputting a portion of the full model vector. The user is not meant to call this function directly.
@@ -29,7 +29,7 @@ class single_transformer(nn.Module):
             self.input_dim = config_dict["num_cosmo_params"] + config_dict["num_nuisance_params"]
         else:
             self.input_dim = config_dict["num_cosmo_params"] + (2 * config_dict["num_nuisance_params"])
-        
+
         self.output_dim = self.num_ells * self.num_kbins
 
         # mlp blocks
@@ -41,24 +41,10 @@ class single_transformer(nn.Module):
                                         self.output_dim,
                                         config_dict["galaxy_ps_emulator"]["num_block_layers"],
                                         config_dict["galaxy_ps_emulator"]["use_skip_connection"]))
-        
-        # expand mlp section output
-        split_dim = config_dict["galaxy_ps_emulator"]["split_dim"]
-        split_size = config_dict["galaxy_ps_emulator"]["split_size"]
-        embedding_dim = split_size*split_dim
-        self.embedding_layer = nn.Linear(self.output_dim, embedding_dim)
 
-        # do one transformer block per z-bin for now
-        self.transformer_blocks = nn.Sequential()
-        for i in range(config_dict["galaxy_ps_emulator"]["num_transformer_blocks"]):
-            self.transformer_blocks.add_module("Transformer"+str(i+1),
-                    blocks.block_transformer_encoder(embedding_dim, split_dim, 0.1))
-            self.transformer_blocks.add_module("Activation"+str(i+1), 
-                    blocks.activation_function(embedding_dim))
+        self.output_layer = nn.Linear(self.output_dim, self.output_dim)
 
-        self.output_layer = nn.Linear(embedding_dim, self.output_dim)
-
-    def forward(self, input_params:torch.Tensor):
+    def forward(self, input_params):
         """Passes an input tensor through the network"""
 
         if not self.is_cross_spectra:
@@ -66,17 +52,15 @@ class single_transformer(nn.Module):
         
         X = self.input_layer(input_params)
         X = self.mlp_blocks(X)
-        X = self.embedding_layer(X)
-        X = self.transformer_blocks(X)
         X = self.output_layer(X)
 
         return X
 
-class stacked_transformer(nn.Module):
-    """Class defining a stack of single_transformer objects, one for each portion of the power spectrum output"""
+class stacked_mlp(nn.Module):
+    """Class defining a stack of single_mlp objects, one for each portion of the power spectrum output"""
 
     def __init__(self, config_dict):
-        """Initializes a group of single_transformer based on the input dictionary.
+        """Initializes a group of single_mlp based on the input dictionary.
         
         This function creates nz*nps total networks, where nz is the number of redshift bins, and nps
         is the number of auto + cross power spectra per redshift bin.
@@ -87,13 +71,13 @@ class stacked_transformer(nn.Module):
         super().__init__()
 
         # output dimensions
-        self.num_zbins = config_dict["num_zbins"]
-        self.num_spectra = config_dict["num_tracers"] +  math.comb(config_dict["num_tracers"], 2)
         self.num_tracers = config_dict["num_tracers"]
+        self.num_spectra = config_dict["num_tracers"] +  math.comb(config_dict["num_tracers"], 2)
+        self.num_zbins = config_dict["num_zbins"]
         self.num_ells = config_dict["num_ells"]
         self.num_kbins = config_dict["num_kbins"]
 
-        self.num_cosmo_params    = config_dict["num_cosmo_params"]
+        self.num_cosmo_params = config_dict["num_cosmo_params"]
         self.num_nuisance_params = config_dict["num_nuisance_params"]
 
         self.output_dim = self.num_ells * self.num_kbins
@@ -103,7 +87,7 @@ class stacked_transformer(nn.Module):
         for z in range(self.num_zbins):
             for isample1, isample2 in itertools.product(range(self.num_tracers), repeat=2):
                 if isample1 > isample2: continue
-                self.networks.append(single_transformer(config_dict, (isample1 != isample2)))
+                self.networks.append(single_mlp(config_dict, (isample1 != isample2)))
 
     def organize_parameters(self, input_params):
         """Organizes input cosmology + bias parameters into a form the rest of the network expects
@@ -147,7 +131,8 @@ class stacked_transformer(nn.Module):
         
         # feed parameters through all sub-networks
         if net_idx == None:
-            X = torch.zeros((input_params.shape[0], self.num_spectra, self.num_zbins, self.output_dim), device=input_params.device)
+            X = torch.zeros((input_params.shape[0], self.num_spectra, self.num_zbins, self.output_dim), 
+                             device=input_params.device)
             
             for (z, ps) in itertools.product(range(self.num_zbins), range(self.num_spectra)):
                 idx = (z * self.num_spectra) + ps
